# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Friendtrip",
        "description": "Friendtrip api document (internal use only)",
        "contact": {
            "email": "chaniwat.meranote@gmail.com"
        },
        "version": "0.0.1"
    },
    "host": "friendtrip.dev",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
        "/authentication": {
            "get": {
                "tags": [
                    "authentication"
                ],
                "summary": "Get authenticate user information",
                "description": "Get user information of given token",
                "operationId": "getAuthUserInfo",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string",
                        "default": "Bearer "
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User information",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "No token provided or invalid"
                    },
                    "401": {
                        "description": "Token expired"
                    }
                }
            },
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "Authentication (Request token)",
                "description": "Request token for user auth needed APIs",
                "operationId": "authentication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User identity",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AuthenticationInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful request token",
                        "schema": {
                            "$ref": "#/definitions/Token"
                        }
                    },
                    "401": {
                        "description": "Invalid credential",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    }
                }
            }
        },
...
        "/users": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Create new user",
                "description": "Create new user into database",
                "operationId": "newUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User information",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/NewUser"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "New user has been created"
                    },
                    "422": {
                        "description": "Invalid parameters"
                    }
                }
            }
        },
        "/users/{user_id}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get user information",
                "description": "Get user information of {user_id} (Need authentication token for some detail)",
                "operationId": "getUserInfo",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "type": "string",
                        "default": "Bearer "
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return user information",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Invalid token"
                    },
                    "401": {
                        "description": "Token expired"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "Update user information",
                "description": "Update user information of {user_id} (Need authentication token for self update or admin token for update to any user)",
                "operationId": "updateUserInfo",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string",
                        "default": "Bearer "
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User information to update",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update user information successful"
                    },
                    "400": {
                        "description": "No token provided or invalid"
                    },
                    "401": {
                        "description": "Token expired or no permission"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/users/{user_id}/password": {
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "Update user password",
                "description": "Update user password of {user_id} (Need authentication token for self update or admin token for update to any user)",
                "operationId": "updateUserPassword",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string",
                        "default": "Bearer "
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "New password",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdatePassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update user password successful"
                    },
                    "400": {
                        "description": "No token provided or invalid"
                    },
                    "401": {
                        "description": "Token expired or no permission"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "AuthenticationInfo": {
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "",
                    "type": "string"
                },
                "password": {
                    "description": "",
                    "type": "string"
                },
                "get_info": {
                    "description": "",
                    "type": "boolean",
                    "default": false
                }
            },
            "type": "object"
        },
...
        "NewUser": {
            "required": [
                "user",
                "password"
            ],
            "properties": {
                "user": {
                    "description": "",
                    "$ref": "#/definitions/UserBody"
                },
                "password": {
                    "description": "",
                    "type": "string"
                }
            },
            "type": "object"
        },
...
        "Token": {
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "description": "",
                    "type": "string"
                },
                "user": {
                    "description": "",
                    "$ref": "#/definitions/User"
                }
            },
            "type": "object"
        },
        "UpdatePassword": {
            "required": [
                "password"
            ],
            "properties": {
                "password": {
                    "description": "",
                    "type": "string"
                }
            },
            "type": "object"
        },
        "User": {
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "",
                    "type": "integer"
                }
            },
            "type": "object",
            "allOf": [
                {
                    "$ref": "#/definitions/UserBody"
                }
            ]
        },
        "UserBody": {
            "required": [
                "email",
                "first_name",
                "last_name",
                "display_name",
                "birthdate",
                "gender",
                "religion"
            ],
            "properties": {
                "email": {
                    "description": "",
                    "type": "string"
                },
                "first_name": {
                    "description": "",
                    "type": "string"
                },
                "last_name": {
                    "description": "",
                    "type": "string"
                },
                "display_name": {
                    "description": "",
                    "type": "string"
                },
                "birthdate": {
                    "description": "",
                    "type": "string",
                    "format": "date"
                },
                "gender": {
                    "description": "",
                    "type": "string",
                    "enum": [
                        "MALE",
                        "FEMALE"
                    ]
                },
                "religion": {
                    "description": "",
                    "type": "string"
                },
                "phone": {
                    "description": "",
                    "type": "string",
                    "maxLength": 10
                }
            },
            "type": "object"
        }
    },
...
}